// berserk-online-client/src/app/store/utils/enums.ts
export enum Elements {
    MOUNTAINS = 'Горы',
    PLAINS = 'Степи',
    FORESTS = 'Леса',
    SWAMPS = 'Болота',
    DARKNESS = 'Тьма',
    NEUTRAL = 'Нейтральная'
}

export enum Rarities {
    COMMON = 'Частая',
    UNCOMMON = 'Необычная',
    RARE = 'Редкая',
    ULTRARARE = 'Ультраредкая'
}

export enum Sets {
    ELEMENTS_WAR = 'Война стихий',
    DARKNESS_INVASION = 'Нашествие тьмы'
}

export enum CardTypes {
    CREATURE = 'creature',
    FLYER = 'flyer',
    COMPANION = 'companion',
    PARASITE = 'parasite',
    ARTIFACT = 'artifact',
    TERRAIN = 'terrain'
}